fix(src): фикс фикса

+++ Content.Server/Imperial/Medieval/Plague/Events/HealingItemUsedEvent.cs
@@ -1,4 +1,4 @@
 namespace Content.Server.Imperial.Medieval.Plague;
 
 [ByRefEvent]
-public record struct PlagueHealingItemUsedEvent(int PlagueDecay);
+public record struct PlagueHealingItemUsedEvent(float PlagueDecay);
